requirements.txt
streamlit==1.36.0
python-dotenv==1.0.1
langchain==0.2.12
langchain-core==0.2.27
langchain-google-genai==1.0.7
langchain-community==0.2.10
google-genai==0.13.0
pydantic==2.11.7
pydantic-core==2.33.2
sqlalchemy==2.0.32
pymysql==1.1.1
pandas==2.2.2
